Moving slot functions back to root
