+++ slots.test.ts
